requirements.txt
openai
google-generativeai